move function convert_to_instrument_pyobject to model crate
